+++ packages/spire/src/index.js
@@ -1,15 +1,10 @@
 //--------------------------------------------------------
 //-- Prettier config
 //--------------------------------------------------------
-import path       from 'path';
-import pkgDir     from 'pkg-dir';
-import fss        from '@absolunet/fss';
 import coreConfig from '@absolunet/prettier-config-core';
+import loader     from '@absolunet/prettier-loader';
 
-module.exports = {
-	...coreConfig,
-	...fss.readYaml(path.join(
-		pkgDir.sync(__dirname),
-		'shareable-config.yaml'
-	))
-};
+module.exports = loader({
+	root:         __dirname,
+	parentConfig: coreConfig
+});